+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ### Added
 - add `PLC_LeakSensor`
 - add `PLC_Fan`
-
+### Changed
+- `PLC_HumidifierDehumidifier` add separated bit on/off support by adding  `set_Deactivate`
+- `PLC_HumidifierDehumidifier` add byte support for rotation speed by adding `set_RotationSpeedByte` and  `get_RotationSpeedByte`
 
 ## [1.0.32] 2021-11-11